# docs/swagger.yaml
basePath: /api/v1
definitions:
  category.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  category.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  category.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  customer.CreateCustomerRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  customer.CustomerResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  customer.UpdateCustomerRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  dashboard.DashboardReportResponse:
    properties:
      product_report:
        $ref: '#/definitions/dashboard.ProductReportResponse'
      top_customers:
        items:
          $ref: '#/definitions/dashboard.TopCustomerResponse'
        type: array
      total_time_ms:
        type: integer
    type: object
  dashboard.ProductReportResponse:
    properties:
      average_price:
        type: number
      cached_at:
        description: Penanda jika data dari cache
        type: string
      recent_products:
        items:
          $ref: '#/definitions/dashboard.RecentProductResponse'
        type: array
      total_products:
        type: integer
      total_stock:
        type: integer
    type: object
  dashboard.RecentProductResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      stock_quantity:
        type: integer
    type: object
  dashboard.TopCustomerResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      total_orders:
        type: integer
      total_spent:
        type: number
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for Order and Dashboard Management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Assignment PTES API
  version: "1.0"
paths:
  /categories:
    get:
      description: Retrieve a list of all categories
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/category.CategoryResponse'
            type: array
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category with name and description
      parameters:
      - description: Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Remove a category from the system by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete category
      tags:
      - categories
    get:
      description: Retrieve a single category by its unique ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.CategoryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update category name or description by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/category.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update category
      tags:
      - categories
  /customers:
    get:
      description: Retrieve a list of all registered customers
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customer.CustomerResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all customers
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Register a new customer with a unique email address
      parameters:
      - description: Customer Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/customer.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
        "400":
          description: Invalid input or email format
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new customer
      tags:
      - customers
  /customers/{id}:
    delete:
      description: Permanently remove a customer from the database
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a customer
      tags:
      - customers
    get:
      description: Retrieve specific customer information by their unique ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
        "404":
          description: Customer not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get customer details
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: Update name or email for an existing customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/customer.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update customer information
      tags:
      - customers
  /dashboard/overview:
    get:
      description: Retrieve a comprehensive report including financial summaries,
        top customers, and product stats
      parameters:
      - description: Limit for sub-lists (top customers/recent items)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dashboard.DashboardReportResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get complete dashboard overview
      tags:
      - dashboard
  /dashboard/products:
    get:
      description: Retrieve summary statistics for products, such as total products,
        active/inactive status, etc.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dashboard.ProductReportResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product dashboard report
      tags:
      - dashboard
  /dashboard/top-customers:
    get:
      description: Retrieve a list of customers with the highest transaction volume
        or spending
      parameters:
      - description: 'Limit the number of customers returned (default: 5)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dashboard.TopCustomerResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get top performing customers
      tags:
      - dashboard
...
swagger: "2.0"